+++ routes/leaderRouter.js
@@ -20,26 +20,49 @@ leaderRouter.route('/')
     .catch((err)=> next(err));
 })
 .post(authenticate.verifyUser,(req,res,next) => {
-    Leaders.create(req.body)
-    .then((leader) => {
-        console.log('Leader Created ',);
-        res.statusCode =200;
-        res.setHeader('Content-Type','application/json');
-        res.json(leader);
-    },(err)=> next(err))
-    .catch((err)=> next(err));
+    if(authenticate.verifyAdmin(req.user)){
+        Leaders.create(req.body)
+        .then((leader) => {
+            console.log('Leader Created ',);
+            res.statusCode =200;
+            res.setHeader('Content-Type','application/json');
+            res.json(leader);
+        },(err)=> next(err))
+        .catch((err)=> next(err));    
+    }
+    else{
+        var err = new Error('You are not authenticated to perform this operation');
+        err.status = 403;
+        next(err);
+    }
+
 })
 .put(authenticate.verifyUser,(req,res,next) => {
-    res.statusCode =403;
-    res.end('PUT operation not supported on /leaders');
+    if(authenticate.verifyAdmin(req.user)){
+        res.statusCode =403;
+        res.end('PUT operation not supported on /leaders');    
+    }
+    else{
+        var err = new Error('You are not authenticated to perform this operation');
+        err.status = 403;
+        next(err);
+    }
+
 })
 .delete(authenticate.verifyUser,(req,res,next)=> {
-    Leaders.remove({})
-    .then((resp)=> {
-        res.statusCode =200;
-        res.setHeader('Content-Type','application/json');
-        res.json(resp);
-    },(err) => next(err))
+    if(authenticate.verifyAdmin(req.user)){
+        Leaders.remove({})
+        .then((resp)=> {
+            res.statusCode =200;
+            res.setHeader('Content-Type','application/json');
+            res.json(resp);
+        },(err) => next(err))    
+    }
+    else{
+        var err = new Error('You are not authenticated to perform this operation');
+        err.status = 403;
+        next(err);
+    }
 });
 
 leaderRouter.route('/:leaderId')
@@ -53,28 +76,52 @@ leaderRouter.route('/:leaderId')
 
 })
 .post(authenticate.verifyUser,(req,res,next)=>{
-    res.statusCode=403;
-    res.end('POST operation not supported on /leaders/'+req.params.leaderId);
+    if(authenticate.verifyAdmin(req.user)){
+        res.statusCode=403;
+        res.end('POST operation not supported on /leaders/'+req.params.leaderId);    
+    }
+    else{
+        var err = new Error('You are not authenticated to perform this operation');
+        err.status = 403;
+        next(err);
+    }
+
 })
 .put(authenticate.verifyUser,(req,res,next)=>{
-    Leaders.findByIdAndUpdate(req.params.leaderId,{
-        $set: req.body
-    },{new:true})
-    .then((leader) => {
-        res.statusCode =200;
-        res.setHeader('Content-Type','application/json');
-        res.json(leader);
-    },(err)=> next(err))
-    .catch((err)=>next(err));
+    if(authenticate.verifyAdmin(req.user)){
+        Leaders.findByIdAndUpdate(req.params.leaderId,{
+            $set: req.body
+        },{new:true})
+        .then((leader) => {
+            res.statusCode =200;
+            res.setHeader('Content-Type','application/json');
+            res.json(leader);
+        },(err)=> next(err))
+        .catch((err)=>next(err));    
+    }
+    else{
+        var err = new Error('You are not authenticated to perform this operation');
+        err.status = 403;
+        next(err);
+    }
+
 })
 .delete(authenticate.verifyUser,(req,res,next)=> {
-    Leaders.findByIdAndRemove(req.params.leaderId)
-    .then((resp)=> {
-        res.statusCode =200;
-        res.setHeader('Content-Type','application/json');
-        res.json(resp);
-    },(err) => next(err))
-    .catch((err)=> next(err));
+    if(authenticate.verifyAdmin(req.user)){
+        Leaders.findByIdAndRemove(req.params.leaderId)
+        .then((resp)=> {
+            res.statusCode =200;
+            res.setHeader('Content-Type','application/json');
+            res.json(resp);
+        },(err) => next(err))
+        .catch((err)=> next(err));    
+    }
+    else{
+        var err = new Error('You are not authenticated to perform this operation');
+        err.status = 403;
+        next(err);
+    }
+
 });
